requirements.txt
Adafruit-Blinka==8.58.1
adafruit-circuitpython-busdevice==5.2.12
adafruit-circuitpython-connectionmanager==3.1.4
adafruit-circuitpython-dht==4.0.8
adafruit-circuitpython-requests==4.1.12
adafruit-circuitpython-typing==1.12.1
Adafruit-PlatformDetect==3.79.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
blinker==1.9.0
click==8.2.1
dht11==0.1.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pyftdi==0.56.0
pyserial==3.5
pyusb==1.3.1
RPi.GPIO==0.7.1
spidev==3.7
sysv_ipc==1.1.0
typing_extensions==4.14.0
Werkzeug==3.1.3
w1thermsensor==1.3.0
